Create 1331_RankTransformOfAnArray.java   Difficulty : Easy
link : https://leetcode.com/problems/rank-transform-of-an-array/description/

description : Given an array of integers arr, replace each element with its rank.
The rank represents how large the element is. The rank has the following rules:
Rank is an integer starting from 1.
The larger the element, the larger the rank. If two elements are equal, their rank must be the same.
Rank should be as small as possible.


Constraints : 
0 <= arr.length <= 10^5
-10^9 <= arr[i] <= 10^9


Topics : 
Array
Hash Table
Sorting
Biweekly Contest 18