Create 1290_ConvertBinaryNumberInALinkedListToInteger.java  Difficulty : Easy
link : https://leetcode.com/problems/convert-binary-number-in-a-linked-list-to-integer/description/

description : Given head which is a reference node to a singly-linked list. The value of each node in the linked list is either 0 or 1. The linked list holds the binary representation of a number.
Return the decimal value of the number in the linked list.
The most significant bit is at the head of the linked list.

Constraints : 
he Linked List is not empty.
Number of nodes will not exceed 30.
Each node's value is either 0 or 1.

Topics : 
Linked List
Math
Weekly Contest 167